+++ crates/matrix-sdk/src/encryption/mod.rs
@@ -1063,7 +1063,7 @@ impl Encryption {
         let olm = self.client.olm_machine().await;
         let olm = olm.as_ref().ok_or(Error::NoOlmMachine)?;
 
-        let keys = olm.export_room_keys(predicate).await?;
+        let keys = olm.store().export_room_keys(predicate).await?;
         let passphrase = zeroize::Zeroizing::new(passphrase.to_owned());
 
         let encrypt = move || -> Result<()> {